+++ packages/imperative/__tests__/plugins/utilities/NpmFunctions.test.ts
@@ -9,16 +9,63 @@
 *
 */
 
+import * as childProcess from "child_process";
 import * as jsonfile from "jsonfile";
 import * as npmPackageArg from "npm-package-arg";
 import * as pacote from "pacote";
-import { getPackageInfo } from "../../../src/plugins/utilities/NpmFunctions";
+import * as npmFunctions from "../../../src/plugins/utilities/NpmFunctions";
 import { PMFConstants } from "../../../src/plugins/utilities/PMFConstants";
 
+jest.mock("child_process");
 jest.mock("jsonfile");
 jest.mock("pacote");
 
 describe("NpmFunctions", () => {
+    const fakeRegistry = "http://localhost:4873/";
+    const npmCmd = npmFunctions.findNpmOnPath();
+
+    afterEach(() => {
+        jest.clearAllMocks();
+    });
+
+    afterAll(() => {
+        jest.restoreAllMocks();
+    });
+
+    it("installPackages should run npm install command", () => {
+        const stdoutBuffer = Buffer.from("Install Succeeded");
+        jest.spyOn(PMFConstants, "instance", "get").mockReturnValueOnce({ PMF_ROOT: __dirname } as any);
+        const spawnSyncSpy = jest.spyOn(childProcess, "spawnSync").mockReturnValueOnce({
+            status: 0,
+            stdout: stdoutBuffer
+        } as any);
+        const result = npmFunctions.installPackages("fakePrefix", fakeRegistry, "samplePlugin");
+        expect(spawnSyncSpy.mock.calls[0][0]).toBe(npmCmd);
+        expect(spawnSyncSpy.mock.calls[0][1]).toEqual(expect.arrayContaining(["install", "samplePlugin"]));
+        expect(spawnSyncSpy.mock.calls[0][1]).toEqual(expect.arrayContaining(["--prefix", "fakePrefix"]));
+        expect(spawnSyncSpy.mock.calls[0][1]).toEqual(expect.arrayContaining(["--registry", fakeRegistry]));
+        expect(result).toBe(stdoutBuffer.toString());
+    });
+
+    it("getRegistry should run npm config command", () => {
+        const stdoutBuffer = Buffer.from(fakeRegistry);
+        const spawnSyncSpy = jest.spyOn(childProcess, "spawnSync").mockReturnValueOnce({
+            status: 0,
+            stdout: stdoutBuffer
+        } as any);
+        const result = npmFunctions.getRegistry();
+        expect(spawnSyncSpy.mock.calls[0][0]).toBe(npmCmd);
+        expect(spawnSyncSpy.mock.calls[0][1]).toEqual(["config", "get", "registry"]);
+        expect(result).toBe(stdoutBuffer.toString());
+    });
+
+    it("npmLogin should run npm login command", () => {
+        const spawnSyncSpy = jest.spyOn(childProcess, "spawnSync").mockReturnValueOnce({ status: 0 } as any);
+        npmFunctions.npmLogin(fakeRegistry);
+        expect(spawnSyncSpy.mock.calls[0][0]).toBe(npmCmd);
+        expect(spawnSyncSpy.mock.calls[0][1]).toContain("login");
+        expect(spawnSyncSpy.mock.calls[0][1]).toEqual(expect.arrayContaining(["--registry", fakeRegistry]));
+    });
 
     describe("getPackageInfo", () => {
         const expectedInfo = { name: "@zowe/imperative", version: "latest" };
@@ -28,14 +75,6 @@ describe("NpmFunctions", () => {
             jest.spyOn(pacote, "manifest").mockResolvedValue(expectedInfo as any);
         });
 
-        afterEach(() => {
-            jest.clearAllMocks();
-        });
-
-        afterAll(() => {
-            jest.restoreAllMocks();
-        });
-
         it("should fetch info for package installed from registry", async () => {
             const pkgSpec = "@zowe/imperative";
             expect(npmPackageArg(pkgSpec).type).toEqual("tag");
@@ -43,7 +82,7 @@ describe("NpmFunctions", () => {
             jest.spyOn(PMFConstants, "instance", "get").mockReturnValueOnce({
                 PLUGIN_NODE_MODULE_LOCATION: ""
             } as any);
-            const actualInfo = await getPackageInfo(pkgSpec);
+            const actualInfo = await npmFunctions.getPackageInfo(pkgSpec);
             expect(actualInfo).toBe(expectedInfo);
             expect(jsonfile.readFileSync).toHaveBeenCalledTimes(1);
         });
@@ -52,7 +91,7 @@ describe("NpmFunctions", () => {
             const pkgSpec = "./imperative";
             expect(npmPackageArg(pkgSpec).type).toEqual("directory");
 
-            const actualInfo = await getPackageInfo(pkgSpec);
+            const actualInfo = await npmFunctions.getPackageInfo(pkgSpec);
             expect(actualInfo).toBe(expectedInfo);
             expect(pacote.manifest).toHaveBeenCalledTimes(1);
         });
@@ -61,7 +100,7 @@ describe("NpmFunctions", () => {
             const pkgSpec = "imperative.tgz";
             expect(npmPackageArg(pkgSpec).type).toEqual("file");
 
-            const actualInfo = await getPackageInfo(pkgSpec);
+            const actualInfo = await npmFunctions.getPackageInfo(pkgSpec);
             expect(actualInfo).toBe(expectedInfo);
             expect(pacote.manifest).toHaveBeenCalledTimes(1);
         });
@@ -70,7 +109,7 @@ describe("NpmFunctions", () => {
             const pkgSpec = "github:zowe/imperative";
             expect(npmPackageArg(pkgSpec).type).toEqual("git");
 
-            const actualInfo = await getPackageInfo(pkgSpec);
+            const actualInfo = await npmFunctions.getPackageInfo(pkgSpec);
             expect(actualInfo).toBe(expectedInfo);
             expect(pacote.manifest).toHaveBeenCalledTimes(1);
         });
@@ -79,7 +118,7 @@ describe("NpmFunctions", () => {
             const pkgSpec = "http://example.com/zowe/imperative.tgz";
             expect(npmPackageArg(pkgSpec).type).toEqual("remote");
 
-            const actualInfo = await getPackageInfo(pkgSpec);
+            const actualInfo = await npmFunctions.getPackageInfo(pkgSpec);
             expect(actualInfo).toBe(expectedInfo);
             expect(pacote.manifest).toHaveBeenCalledTimes(1);
         });