+++ packages/imperative/__tests__/plugins/utilities/npm-interface/uninstall.test.ts
@@ -27,7 +27,7 @@ import { IPluginJson } from "../../../../src/plugins/doc/IPluginJson";
 import { Logger } from "../../../../../logger";
 import { PMFConstants } from "../../../../src/plugins/utilities/PMFConstants";
 import { readFileSync, writeFileSync } from "jsonfile";
-import { cmdToRun } from "../../../../src/plugins/utilities/NpmFunctions";
+import { findNpmOnPath } from "../../../../src/plugins/utilities/NpmFunctions";
 import { uninstall } from "../../../../src/plugins/utilities/npm-interface";
 
 
@@ -42,7 +42,7 @@ describe("PMF: Uninstall Interface", () => {
     const samplePackageName = "imperative-sample-plugin";
     const packageName = "a";
     const packageRegistry = "https://registry.npmjs.org/";
-    const npmCmd = cmdToRun();
+    const npmCmd = findNpmOnPath();
 
     beforeEach(() => {
     // Mocks need cleared after every test for clean test runs
@@ -101,7 +101,10 @@ describe("PMF: Uninstall Interface", () => {
 
     describe("Basic uninstall", () => {
         beforeEach(() => {
-            mocks.spawnSync.mockReturnValue(`+ ${packageName}`);
+            mocks.spawnSync.mockReturnValue({
+                status: 0,
+                stdout: Buffer.from(`+ ${packageName}`)
+            } as any);
         });
 
         it("should uninstall", () => {